+++ CHANGELOG.md
@@ -24,6 +24,8 @@ Change categories:
 - utils: Added `pldm_edac_crc32_extend()`
 - base: Added `decode_pldm_base_negotiate_transfer_params_req()`
 - base: Added `encode_pldm_base_negotiate_transfer_params_resp()`
+- pdr: Add foreach_pldm_pdr_effecter_possible_states() iterator macro
+- pdr: Add foreach_pldm_pdr_effecter_bitfield_states() iterator macro
 
 ### Changed